requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
Pillow==10.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1